scripts/train_ocr.py
"""
Fine-tune PaddleOCR for medical document recognition
"""

from paddleocr import PaddleOCR
from loguru import logger
import yaml


def load_config(config_path='config/training.yaml'):
    """Load training configuration"""
    with open(config_path, 'r') as f:
        return yaml.safe_load(f)


def prepare_dataset(data_dir):
    """
    Prepare medical document dataset
    Expected structure:
    data/
      train/
        images/
        labels.txt
      val/
        images/
        labels.txt
    """
    logger.info(f"Loading dataset from {data_dir}")
    # Dataset preparation logic here
    return {
        'train_samples': 1000,
        'val_samples': 200
    }


def fine_tune_ocr(config):
    """Fine-tune PaddleOCR model"""
    logger.info("Starting OCR fine-tuning")
    
    # Training parameters
    params = {
        'model': 'paddleocr-vl',
        'learning_rate': 2e-5,
        'batch_size': 16,
        'epochs': 20,
        'optimizer': 'AdamW'
    }
    
    logger.info(f"Training params: {params}")
    
    # Placeholder for actual training
    # In production, would use PaddleOCR training API
    
    logger.info("Training complete")
    logger.info("Model saved to: models/paddleocr-medical")


if __name__ == "__main__":
    config = load_config()
    dataset_info = prepare_dataset('data/medical_docs')
    fine_tune_ocr(config)
    
    logger.info("OCR fine-tuning pipeline complete")
