requirements.txt
# Core dependencies
paddlepaddle-gpu==2.6.0
paddleocr==2.7.3
erniebot==1.0.0
camel-ai==0.1.5.5

# ML/AI frameworks
torch==2.1.0
transformers==4.35.0
llama-factory==0.3.1

# Web framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Data processing
pandas==2.1.3
numpy==1.24.3
pillow==10.1.0
opencv-python==4.8.1.78

# API clients
requests==2.31.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.23
pymongo==4.6.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Edge deployment
onnxruntime==1.16.3
openvino==2023.2.0

# Cloud services
novita-client==0.5.2
baidu-aip==4.16.13
